+++ src/filter/fir_no_simd_impl.cpp
@@ -7,7 +7,7 @@ template class FirNoSimdImpl<Complex>;
 
 
 template<>
-FirNoSimdImpl<Real>::FirNoSimdImpl(uint32_t sample_rate, const std::vector<Real>& coeff) :
+FirNoSimdImpl<Real>::FirNoSimdImpl(double sample_rate, const std::vector<Real>& coeff) :
 	Filter(sample_rate, std::vector<Filter::FilterOutputType>(1, Filter::FilterOutputType::REAL)),
 	coefficients(coeff.crbegin(), coeff.crend()),
 	buffer_real(coeff.size(), Real(0))
@@ -17,7 +17,7 @@ FirNoSimdImpl<Real>::FirNoSimdImpl(uint32_t sample_rate, const std::vector<Real>
 
 
 template<>
-FirNoSimdImpl<Complex>::FirNoSimdImpl(uint32_t sample_rate, const std::vector<Real>& coeff) :
+FirNoSimdImpl<Complex>::FirNoSimdImpl(double sample_rate, const std::vector<Real>& coeff) :
 	Filter(sample_rate, std::vector<Filter::FilterOutputType>(1, Filter::FilterOutputType::COMPLEX)),
 	coefficients(coeff.crbegin(), coeff.crend()),
 	buffer_real(coeff.size(), Real(0)),
